fixed bug where numbers were not updating properly
lack of reset in the calculate function caused it so occasionally when
entering multiple calculations, the calculator would stop working as
intended. Changed the way the variables are updated and made sure
calculate resetted both the operator and the second number. Also made it
so the operation functions can call calculate()